+++ src/main/kotlin/io/github/inductiveautomation/kindling/core/Tool.kt
@@ -7,7 +7,7 @@ import io.github.inductiveautomation.kindling.gatewaynetwork.GatewayNetworkTool
 import io.github.inductiveautomation.kindling.idb.IdbViewer
 import io.github.inductiveautomation.kindling.localization.TranslationTool
 import io.github.inductiveautomation.kindling.log.LogViewer
-import io.github.inductiveautomation.kindling.questdb.QuestdbViewer
+import io.github.inductiveautomation.kindling.questdb.QuestDbViewer
 import io.github.inductiveautomation.kindling.serial.SerialViewer
 import io.github.inductiveautomation.kindling.thread.MultiThreadViewer
 import io.github.inductiveautomation.kindling.utils.FileFilter
@@ -58,7 +58,7 @@ interface Tool : KindlingSerializable {
                 MultiThreadViewer,
                 LogViewer,
                 IdbViewer,
-                QuestdbViewer,
+                QuestDbViewer,
                 CacheViewer,
                 GatewayNetworkTool,
                 AlarmViewer,
