+++ src/main/kotlin/io/github/inductiveautomation/kindling/core/db/Table.kt
@@ -1,4 +1,4 @@
-package io.github.inductiveautomation.kindling.idb.generic
+package io.github.inductiveautomation.kindling.core.db
 
 import java.util.Collections
 import java.util.Enumeration
